pyproject.toml
[project]
name = "sec-review-framework"
version = "0.1.0"
description = "AI/LLM Security Code Review Testing Framework"
requires-python = ">=3.14"
dependencies = [
    "pydantic>=2.10",
    "pyyaml>=6.0",
    "litellm>=1.60",
    "httpx>=0.28",
    "numpy>=2.2",
    "scipy>=1.15",
]

[project.optional-dependencies]
coordinator = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.34",
    "kubernetes>=32.0",
    "prometheus-client>=0.22",
    "aiosqlite>=0.21",
    "python-multipart>=0.0.18",
]
worker = [
    "semgrep>=1.100",
]
dev = [
    "pytest>=8.3",
    "pytest-asyncio>=0.25",
    "ruff>=0.11",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/sec_review_framework"]

[tool.ruff]
target-version = "py314"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: requires temp database or mock services",
    "infra: requires Docker or K8s tools",
]

[dependency-groups]
dev = [
    "pytest-cov>=7.1.0",
]
